drm/i915: Fix kernel crash after GPU HANG
While using execlists without any display attached, there is a kernel crash
observed with intel joule compute module (gen9 gpu).

The crash happens while resetting the gpu, in reset_common_ring(), where
port[0].request can be NULL.

This doesn't solve the actual GPU HANG issue leading to this situation, but
it is not acceptable to crash the whole kernel for it.

A snippet of the crash dump:

[   84.902889] drm/i915: Resetting chip after gpu hang
[   84.910547] BUG: unable to handle kernel NULL pointer dereference at 0000000000000070
[   84.919385] IP: [<ffffffffa022e4c2>] reset_common_ring+0xa2/0x130 [i915]
[   84.926941] PGD 0 [   84.928998]
[   84.930670] Oops: 0000 [#1] PREEMPT SMP

Signed-off-by: Jukka Laitinen <jukka.laitinen@intel.com>